car-video-client/requirements.txt
aiohttp==3.9.5
aioice==0.10.2
aiortc==1.6.0
aiosignal==1.4.0
attrs==25.4.0
av==11.0.0
cffi==2.0.0
cryptography==46.0.3
dnspython==2.8.0
frozenlist==1.8.0
google-crc32c==1.8.0
idna==3.11
ifaddr==0.2.0
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
pillow==10.3.0
propcache==0.4.1
pycparser==2.23
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
python-dotenv==1.0.1
typing_extensions==4.15.0
websockets==12.0
yarl==1.22.0
